fix patch file
